Split client interfacing code into FileProviderExtension extension
