+++ shell_integration/MacOSX/NextcloudIntegration/FileProviderExt/FileProviderExtension.swift
@@ -630,54 +630,4 @@ class FileProviderExtension: NSObject, NSFileProviderReplicatedExtension, NKComm
 
         fpManager.signalEnumerator(for: .workingSet, completionHandler: completionHandler)
     }
-
-    // MARK: Nextcloud desktop client communication
-    func sendFileProviderDomainIdentifier() {
-        let command = "FILE_PROVIDER_DOMAIN_IDENTIFIER_REQUEST_REPLY"
-        let argument = domain.identifier.rawValue
-        let message = command + ":" + argument + "\n"
-        socketClient?.sendMessage(message)
-    }
-
-    private func signalEnumeratorAfterAccountSetup() {
-        guard let fpManager = NSFileProviderManager(for: domain) else {
-            Logger.fileProviderExtension.error("Could not get file provider manager for domain \(self.domain.displayName, privacy: OSLogPrivacy.auto(mask: .hash)), cannot notify after account setup")
-            return
-        }
-
-        assert(ncAccount != nil)
-
-        fpManager.signalErrorResolved(NSFileProviderError(.notAuthenticated)) { error in
-            if error != nil {
-                Logger.fileProviderExtension.error("Error resolving not authenticated, received error: \(error!.localizedDescription)")
-            }
-        }
-
-        Logger.fileProviderExtension.debug("Signalling enumerators for user \(self.ncAccount!.username) at server \(self.ncAccount!.serverUrl, privacy: OSLogPrivacy.auto(mask: .hash))")
-        fpManager.signalEnumerator(for: .workingSet) { error in
-            if error != nil {
-                Logger.fileProviderExtension.error("Error signalling enumerator for working set, received error: \(error!.localizedDescription, privacy: .public)")
-            }
-        }
-    }
-
-    func setupDomainAccount(user: String, serverUrl: String, password: String) {
-        ncAccount = NextcloudAccount(user: user, serverUrl: serverUrl, password: password)
-        ncKit.setup(user: ncAccount!.username,
-                     userId: ncAccount!.username,
-                     password: ncAccount!.password,
-                     urlBase: ncAccount!.serverUrl,
-                     userAgent: "Nextcloud-macOS/FileProviderExt",
-                     nextcloudVersion: 25,
-                     delegate: nil) // TODO: add delegate methods for self
-
-        Logger.fileProviderExtension.info("Nextcloud account set up in File Provider extension for user: \(user, privacy: OSLogPrivacy.auto(mask: .hash)) at server: \(serverUrl, privacy: OSLogPrivacy.auto(mask: .hash))")
-
-        signalEnumeratorAfterAccountSetup()
-    }
-
-    func removeAccountConfig() {
-        Logger.fileProviderExtension.info("Received instruction to remove account data for user \(self.ncAccount!.username, privacy: OSLogPrivacy.auto(mask: .hash)) at server \(self.ncAccount!.serverUrl, privacy: OSLogPrivacy.auto(mask: .hash))")
-        ncAccount = nil
-    }
 }
